refactor: Move pages to individual components.
Refactor the Settings screen on both Android and iOS by decomposing monolithic screen files into smaller, page-specific components. This improves maintainability and readability by separating the root settings list from individual configuration pages.

*   **Android (Compose):**
    *   Decompose `SettingsScreen.kt` by extracting sub-pages into a dedicated `components` package.
    *   Create standalone components for `AppearancePage`, `BehaviorPage`, `InfoPage`, `LicensesPage`, `NotificationsPage`, `PrivacyPage`, `SettingsRootContent`, and `TraktPage`.
    *   Relocate browser-based URL launching logic to a reusable `openInCustomTab` helper in `SettingsComponents.kt`.
    *   Update Trakt connection descriptions to be more concise and user-friendly.

*   **iOS (SwiftUI):**
    *   Refactor the `Settings` package to use a modular component architecture, extracting UI logic from `SettingsScreen.swift`.
    *   Introduce page-specific views: `AppearancePageView`, `InfoPageView`, `LicensesPageView`, `PrivacyPageView`, `SettingsRootContentView`, `SettingsTogglesPageView`, and `TraktPageView`.
    *   Abstract common UI patterns into reusable building blocks within `SettingsComponents.swift`, including `SettingsCard`, `SettingsNavigationRow`, `SettingsToggleRow`, and `SettingsLinkRow`.
    *   Define dedicated data structures for settings pages (e.g., `SettingsInfoContent`, `SettingsTraktContent`) in `SettingsModels.swift`.
    *   Add `SettingsPreviewSamples` to provide robust mock data for SwiftUI Previews.

*   **General:**
    *   Clean up redundant constants, imports, and private helper methods from the primary screen entry points.
    *   Update Android UI snapshot tests to reflect structural changes and label updates.
