+++ ios/Packages/Settings/Sources/Settings/Components/AppearancePageView.swift
@@ -1,0 +1,62 @@
+import Components
+import DesignSystem
+import SwiftUI
+
+struct AppearancePageView<Theme: ThemeItem>: View {
+    @Environment(\.appTheme) private var appTheme
+    private let themeItem: SettingsThemeItem<Theme>
+    private let imageQualityItem: SettingsImageQualityItem
+
+    init(themeItem: SettingsThemeItem<Theme>, imageQualityItem: SettingsImageQualityItem) {
+        self.themeItem = themeItem
+        self.imageQualityItem = imageQualityItem
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: appTheme.spacing.large) {
+            VStack(alignment: .leading, spacing: appTheme.spacing.small) {
+                SettingsSectionLabel(themeItem.title)
+                SettingsCard {
+                    ThemeSelectorView(
+                        themes: themeItem.themes,
+                        selectedTheme: themeItem.selectedTheme,
+                        onThemeSelected: themeItem.onThemeSelected
+                    )
+                    .padding(appTheme.spacing.medium)
+                }
+            }
+
+            VStack(alignment: .leading, spacing: appTheme.spacing.small) {
+                SettingsSectionLabel(imageQualityItem.title)
+                SettingsCard {
+                    VStack(alignment: .leading, spacing: appTheme.spacing.small) {
+                        HStack(spacing: appTheme.spacing.small) {
+                            ForEach(imageQualityItem.options) { option in
+                                SelectionChip(
+                                    label: option.label,
+                                    isSelected: option.id == imageQualityItem.selectedOptionId,
+                                    action: option.onSelect
+                                )
+                            }
+                        }
+                        Text(imageQualityItem.subtitle)
+                            .textStyle(appTheme.typography.bodySmall)
+                            .foregroundColor(appTheme.colors.onSurfaceVariant)
+                    }
+                    .padding(appTheme.spacing.medium)
+                }
+            }
+        }
+    }
+}
+
+#if DEBUG
+#Preview {
+    AppearancePageView(
+        themeItem: SettingsPreviewSamples.themeItem,
+        imageQualityItem: SettingsPreviewSamples.imageQualityItem
+    )
+    .padding()
+    .appPreview()
+}
+#endif
